+++ DKPlugins/DKCurl/DKCurl.h
@@ -59,7 +59,8 @@ public:
 	bool HttpFileSize(const DKString& url, long& size);
 	bool HttpToString(const DKString& url, DKString& output);
 
-	static int WriteToBuffer(char *data, size_t size, size_t nmemb, std::string *buffer);
+	//static int WriteToBuffer(char *data, size_t size, size_t nmemb, std::string *buffer);
+	static size_t WriteToBuffer(char* data, size_t size, size_t nmemb, std::string* buffer);
 	static size_t WriteToFile(void *ptr, size_t size, size_t nmemb, FILE *stream);
 	static size_t read_callback(void *ptr, size_t size, size_t nmemb, void *stream);
 	static int progress_func(void* ptr, double TotalToDownload, double NowDownloaded, double TotalToUpload, double NowUploaded);
